+++ pip/models/__init__.py
@@ -1,0 +1,1 @@
+from pip.models.index import Index, PyPI  # noqa
